Modify med5 sequence to match the actual modeling.
A typo in the original modeling script caused the med5 subunit
to be extended by 14 residues. In order for our current
models to match those published, we need to add those
extra residues to the med5 sequence itself. This shouldn't
affect the models that much since the number of beads
representing the C terminal region of med5 doesn't change
either way. Relates #2.