Add simple script to split the repo into zip files.
We do this to make it easier to access the data from Zenodo.
Putting the entire repository in a single zip file means users
have to download an enormous file to see any of the data, but
in most cases people aren't interested in trajectories and other
large parts of the repository.